move near to subdir
