+++ packages/advanced-logic/src/extensions/payment-network/near/any-to-near.ts
@@ -1,6 +1,6 @@
 import { ICurrencyManager, UnsupportedCurrencyError } from '@requestnetwork/currency';
 import { ExtensionTypes, IdentityTypes, RequestLogicTypes } from '@requestnetwork/types';
-import AnyToNativeTokenPaymentNetwork from './any-to-native';
+import AnyToNativeTokenPaymentNetwork from '../any-to-native';
 
 const CURRENT_VERSION = '0.1.0';
 const NETWORK = 'aurora';
