+++ packages/advanced-logic/test/extensions/payment-network/any-to-near.test.ts
@@ -9,11 +9,11 @@ import {
 import { AdvancedLogic } from '../../../src';
 import { arbitraryTimestamp, payeeRaw, payerRaw } from '../../utils/test-data-generator';
 import { ExtensionTypes, RequestLogicTypes } from '@requestnetwork/types';
-import AnyToNearPaymentNetwork from '../../../src/extensions/payment-network/any-to-near';
+import AnyToNearPaymentNetwork from '../../../src/extensions/payment-network/near/any-to-near';
 import AnyToNativeTokenPaymentNetwork from '../../../src/extensions/payment-network/any-to-native';
 import { CurrencyManager } from '@requestnetwork/currency';
 import utils from '@requestnetwork/utils';
-import AnyToNearTestnetPaymentNetwork from '../../../src/extensions/payment-network/any-to-near-testnet';
+import AnyToNearTestnetPaymentNetwork from '../../../src/extensions/payment-network/near/any-to-near-testnet';
 
 const salt = arbitrarySalt;
 const currencyManager = CurrencyManager.getDefault();
